requirements.txt
fastapi==0.88.0
flask-sqlalchemy==3.0.2
prometheus_client==0.15.0
confluent-kafka==1.9.2
ujson==5.6.0
APScheduler==3.9.1
jira==3.4.1
skpy==0.10.4
requests-cache==0.9.7
retry_decorator==1.1.1
python-logging-loki==0.3.1
opentelemetry-instrumentation-logging==0.36b0
opentelemetry-exporter-otlp==1.14.0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-instrumentation-confluent-kafka==0.36b0
prometheus_fastapi_instrumentator==5.9.1
pytz==2022.7
uvicorn==0.20.0
gunicorn==20.1.0
SQLAlchemy==1.4.46
mariadb==1.0.11